requirements-current.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alabaster==1.0.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
Authlib==1.6.0
babel==2.17.0
backoff==2.2.1
bandit==1.8.5
bcrypt==4.3.0
black==25.1.0
boolean.py==5.0
build==1.2.2.post1
bump2version==1.0.1
CacheControl==0.14.3
cachetools==5.5.2
cattrs==25.1.1
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
coverage==7.9.1
cryptography==45.0.4
customtkinter==5.2.2
cycler==0.12.1
cyclonedx-python-lib==9.1.0
darkdetect==0.8.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
docstring-to-markdown==0.17
docutils==0.21.2
dparse==0.6.4
durationpy==0.10
eval_type_backport==0.2.2
fastapi==0.115.9
filelock==3.16.1
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.28.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
identify==2.6.12
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
interrogate==1.7.0
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jedi-language-server==0.45.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
kiwisolver==1.4.8
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
license-expression==30.4.1
line_profiler==4.2.0
lsprotocol==2023.0.1
macholib==1.16.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mistralai==1.8.2
mmh3==5.1.0
modulegraph==0.19.6
more-itertools==10.7.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
mypy==1.16.1
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.0
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.88.0
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
osascript==2020.12.3
overrides==7.7.0
packageurl-python==0.17.1
packaging==24.2
pandas==2.3.0
parso==0.8.4
pathspec==0.12.1
pbr==6.1.1
pillow==11.2.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
posthog==5.0.0
pre_commit==4.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
py==1.11.0
py-serializable==2.0.0
py-spy==0.4.0
py2app==0.28.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
pydocstyle==6.3.0
pyflakes==3.4.0
PyGetWindow==0.0.9
pygls==1.3.1
Pygments==2.19.2
PyMsgBox==1.0.9
pynput==1.8.1
pyobjc-core==11.1
pyobjc-framework-ApplicationServices==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreServices==11.1
pyobjc-framework-CoreText==11.1
pyobjc-framework-FSEvents==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
pyperclip==1.9.0
PyPika==0.48.9
pyproject-api==1.9.0
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.4.0
pytest-asyncio==1.0.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
roman-numerals-py==3.1.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
rubicon-objc==0.5.1
ruff==0.12.0
rumps==0.4.0
runcmd==2020.12.3
safetensors==0.5.3
safety==3.5.2
safety-schemas==0.0.14
scalene==1.5.51
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
starlette==0.45.3
stevedore==5.4.1
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tkinter-tooltip==3.1.2
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
tox==4.27.0
tqdm==4.67.1
transformers==4.52.4
Tree==0.2.4
typer==0.16.0
types-PyYAML==6.0.12.20250516
types-requests==2.32.4.20250611
types-setuptools==80.9.0.20250529
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
